requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
flask-marshmallow==0.10.1
uwsgi==2.0.18
marshmallow-sqlalchemy==0.17.1
Flask-Migrate==2.5.2
